saperated files functionality
