+++ .gitignore
@@ -1,3 +1,3 @@
 venv/
 .idea
-*.pyc
+**/__pycache__/
