+++ datastructures.py
@@ -1,0 +1,5 @@
+class WeatherData:
+    @classmethod
+    def __init__(self, filename, day_data):
+        self.filename = filename
+        self.day_data = day_data
